server/requirements.txt
Flask
Gunicorn
google-cloud-language>=2.0.0,<3.0.0
flask_cors
firebase_admin